+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Project Analysis feature added showing Javascript and Asset sizes
 - Capacitor 3 to 4 migration feature added
+- Recommend install of @awesome-cordova-plugins/core if using a plugin
+- Recommend @capacitor/google-maps on Capacitor migration where cordova-plugin-googlemaps is present
 - Add generated splash and icon resources to .gitignore when generating for Capacitor
 - Fix for retrying builds when a bug is fixed in code and file saved